feat: refactor pdf util to separate utils file
